Return no value from %SET-SAP-REF-* vops
In addition to removing the result operand, remove all extra complications
for the 32-bit x86 code. The maintenance burden was too high to continue
having different vops for just the one architecture.

Fixes the usual problem of consing if the result was used.
Also there was a bug in sap-ref-64 for 32-bit word size where the
compiler transform accidentally chopped the result, which this fixes.