Avoid accessing all_threads without locks.
Before a thread is initialized it's filled with no-tls-marker, on CPUs
with relaxed memory ordering the marker may still remain there after
the thread is pushed to all_threads.

sprof.c calls gc_managed_addr_p, which loops through all_threads, but
it doesn't actually need to examine them, gc_managed_heap_space_p is
enough.

Most of the other uses of all_threads are during GCing, with the
all_threads_lock grabbed.

(Would filling with 0 and not 97 be a better idea? Should all uses of
all_threads be audited?)