block compilation: Fix issues with DEFPACKAGE and DEFCONSTANT.
Fopcompilation is not the way to solve these issues, as it poses
issues with evaluation ordering semantics and cannot generally handle
DEFCONSTANT with non-fopcompilable forms. (see the added and the
adfixed test) Instead, we must resort to having these forms delimit
block compilation, solving both the fopcompilable issue as well as the
evaluation ordering issue.

A quick proof that this is not only sufficient to solve the bug but
that this fix is necessary and maximal in terms of compiling as much
code together as possible in general circumstances:

Suppose TLF-n is a form that defines a constant or package X
referenced by later TLFs. Then loading components that reference X
must have access to X at load time. Consider a file with the
following:

TLF-1 etc.
TLF-n
TLF-n+1 etc.

Because of evaluation order constraints, TLF-k-1 must run before
TLF-k. The component(s) containing non-top-level code in TLF-k must
also be loaded prior to TLF-k's execution in general.

The code for TLF-n+1 etc. may reference X, so it must be in a separate
component as TLF-n for fasl loading to work properly. In addition,
TLF-n+1 etc must run after TLF-n.

Hence, we have shown that there must be at least two code components
in this situation and that none of the code in TLF-1 up to TLF-n can
share a component with code in the rest of the TLFs.

Proof end.

It may be the case that more code can be compiled together in cases
where the above constraints in the proof assumptions do not hold, but
it does not seem feasible/worthwhile to detect those cases.