Optionally add more assertions if #+bignum-assertions
This change finds places that bignum logic assumes zero-fill in the allocator
and makes them not assume that. Eliminating the assumption enables allocating
most if not all unboxed objects to pages that aren't prezeroed, which will
reduce the sizeable chunk of time that the C runtime spends doing memset().

Tested on 32-bit ppc, x86, and x86-64