Demonstrate some ways to improve WITH-MUTEX
We can get at least 25% speedup by eschewing all the async unwind crap.
There's also some improvement possible by using the mutex3 algorithm
from the "Futexes are Tricky" paper. These tests also show that it is
theoretically possible to retain synchronous NLX safety by chaining all
owned mutexes onto a dynamic-extent list; when unwinding the binding
stack due to NLX we would just need to release each held mutex.

As to async unwind safety, it is mostly unreliable garbage anyway,
and nobody should need it in production code. Performance would presumably
not be a concern for a subset of users. It's just a question of how to
convert from the old API to a new one that does not promise asynchronous
interrupt safety where users have come to expect it.