Don't instrument %DEFSTRUCT for cold load.
First of all, this seems to have been purely a debugging aid. Second
of all, any DEFSTRUCTs which are block compiled will not get
instrumented because this instrumentation relies on the fopcompiler
which is mutually exclusive wiht block compilation. Therefore, it's
better not to do this anymore since it's confusing that it only works
sometimes.

There's also a comment which says we might be able to do this another
way, in case the need arises to do this again.