Improve logbitp on fixnum and raw structure slots
As with logtest, absorb the memory load into the instruction.
When we can absorb the load, then the instruction can always
be emitted as a TEST of 1 bit, picking out the desired bit.
Actually in all cases try to emit as TEST rather than BT, except
when TEST would nede a huge constant operand such as (ASH 1 60).

Also fix lp#1939897