Skip 0-filling in %shrink-vector
Turn the object that follows the shrunken vector into a filler,
which takes exactly 1 store and is therefore atomic.
But if GC should happen to see the old length after the new object
header is written (because it has to be written first),
there is really no harm whatsoever in that.