Allow changing SIG_STOP_FOR_GC to something else
We shouldn't need this change, because SIGUSR2 is conventionally reserved for
applications (and not libraries) - and as another data point, OpenJDK uses it -
but unfortunately LLVM uses it too, and it's sort of David vs. Goliath,
with SBCL being David. Unlike in the story, however, Goliath wins.

So now, if SIGUSR2 is unworkable for you, then you can pick whatever you like
as long as what you like is either SIGVTALRM or SIGWINCH.