Don't create casts in substitute-single-use-lvar.
If it happens at the end of the alloted optimization passes the
created cast may never see IR1-OPTIMIZE-CAST, which sets
CAST-%TYPE-CHECK to NIL, without which checkgen fails.
Could set CAST-%TYPE-CHECK to NIL in SUBSTITUTE-SINGLE-USE-LVAR, but
since its only goal is to derive the type from the ref to its
replacement arg's uses, do that directly.

Fixes lp#1943327