Fix busy-looping on win32 sockets
On windows, HANDLE-LISTEN busy-loops causing the CPU to shoot up the
roof when waiting on input from file descriptors (sockets, pipes, and
consoles) using WAIT-UNTIL-FD-USABLE.

Expanding HANDLE-LISTEN to receive two optional arguments, allows us
to keep the old interface and to pass the timeouts down to the SELECT
happening on the socket.

Future work should try and make the same for pipes and consoles.