Change primtype rules for SIMD packs.
Before this commit, objects there were known to be SIMD packs, but whose
element type was not clear, were still assigned a primitive type.  This
allows writing VOPs that ignore the element type, but also has some nasty
consequences.  One consequence is that type dispatch on SIMD packs was
broken under some circumstances, such that, e.g., PRINT-OBJECT used the
wrong printed representation for some SIMD packs.  The other problem is
that VOPs and function that ignore the SIMD pack element type undermine
SBCL's ability to do static type checking and warn about things like a SIMD
pack of single floats being passed to a function that expects a SIMD pack
of double floats.

This commit fixes PRINT-OBJECT on SIMD packs, and enforces a strict SIMD
programming style where in efficient code, each pack must be known to be of
a particular element type.  If a programmer actually wants to reinterpret a
SIMD pack with a different element type, we expect them to use explicit
casts.  Such casts will be introduced by the sb-simd contrib, soon.