Make make-host-1 quiet again
I had put great effort into linearizing the build so that our datastructures
were presented to the compiler in as acyclic a way as possible.
All that effort was undone, so the host complains loudly again.
Of course I hate this change, but I hate reams of useless output more.