clean up and fix memory leak in search_for_core()
Changed allocation to successful_malloc instead of calloc
Use snprintf instead of sprintf
Removed duplicate and unnecessary free(), since the newly allocated buffer was smaller anyway.
A specific case for returning NULL is unneeded, since it's the same as returning core itself in that case.
The buffer lookhere wasn't freed in the old return NULL case.