x86-64: remove lots more SSE disassembly formats
Extract the optional REX byte prior to calling FIND-INST
so that we don't have 2x as many patterns to match.

Something similar could work with VEX2 versus VEX3 because it should
be possible to recast any instruction that uses VEX2 as if it used VEX3
in the pre-decode function by returning a modified byte sequence,
again cutting the number of patterns in half.