riscv64: Fix complex single float.
The ISA specification now defines the behavior of extracting single
floats from double float registers more precisely. In particular, we
have to be careful not to reinterpret a double as a single directly,
since the spec now mandates that only NaN-boxed values can be
interpreted directly as singles. Anything else is subject to become
a canonical NaN.