Handle zero-tn on arm64 differently.
Some other backends always put 0 into a zero-sc constant, but not all
VOPs want to use zr-tn, some want to use immediate values. Define
zero-sc but don't plug it in any of the load functions, instead, after
select-representations is finished change all the references of the 0
constant for the VOPs that accept zero-sc to a wired zero TN.

This simplifies the previous load-if expressions and doesn't allocate a
load-tn.