Take out a failing dcheck
It seems not to matter. I noticed it failing quite reliably in info.impure
while trying track down the (unrelated) reason for failure of the
assertion in scav_code_header() that a page is not protected.
I still don't understand that one, and I'm having trouble with it
because in order to remove all use of signal-based mprotect() on the
dynamic-spcae, it's essential to find all places where manual toggling
of the logical protection bit has to be performed.