Move the write_protected bit out of page_table
All machines that we compile multithreaded SBCL for can store to a byte
without affecting other bytes in a word and without needing an atomic
operation per se. (At least, that's my understanding - it doesn't cause
bits in other bytes of the word to become randomized)
This will be important when explicitly emitting write barriers from lisp,
but meanwhile it eliminates the dependency on sizeof (struct page)
which simplifies the CODE-HEADER-SET asm routines.