Fix long-standing bug in re-applying write-protect status
This logic has always been both suboptimal and wrong, potentially erring
in either direction: not WPing when it could, and WPing when it mustn't.
Only the latter is an actual bug, and nearly impossible to observe.

One bug was already documented as pertains to FDEFNs. This seems never
to happen, so I'm leaving it alone for now. The observable bug was with
lockfree lists where an element is pending pending physical deletion
after being logically deleted, with GC occurring in between those steps.

The suboptimalities are several (but I'm not fixing them all):
- deposit_filler() does not zeroize words that were formerly occupied
  by an object that was evacuated from the partially pinned pages.
  There were frequently random pointers to lower-generation pages.
- unboxed objects may reside on boxed pages if we were unable to move
  them because of pinning, so we'd treat raw bits as pointers. Whether
  they look like pointers is about as rare as the situation that a stack
  word falsely looks like a pointer, but it's easy enough to skip over
  unboxed objects, same as with filler objects.
- mixed raw/boxed instances should use the bitmap.  It's unclear whether
  doing so would be faster. Treating every word as a pointer is fine
  but continues to be slightly suboptimal.