ppc: Fix crashes in (TRACE :ENCAPSULATE NIL) tests
They began failing after I added a %CODE-FIXUPS slot, but I didn't notice.
It doesn't help that we allow known test suite failures for ppc which
are not annotated as such, because you have to use your memory to decide
whether each test is newly failing or not.

So I just plugged in random values of N for CODE_SIZE+N until it didn't crash,
because I'm not going to pretend to understand that logic any more.
Obviously the comment about it _not_ being required to adjust reg_CODE
properly is incorrect, so remove it.