ppc64: implement soft card marks
Register NL5 (GPR 8) is dedicated as the card table base. It has to be
restored after each call into C because it's volatile across calls.
But foreign calls are so horrifically bad anyway based on the amount of
required boilerplate that one load is insigificant, and a wired register
reduces the GC barrier to just 2 instructions per site.
Without reserving a register, this change was performance-neutral,
but with NL5 wired, it resulted in about a 1% speedup.

Anyway this change shows how simple it is to modify the vops.
Hopefully it will serve as a model for other architectures to follow.