Do not inhibit block joining for EXIT ctrans.
I'm not positive this case can happen anymore, now that translation of
BLOCK and TAGBODY have been simplified. Certainly, the test suite now
passes without this. However, even if something like this does need to
be reinstated, it would be better if block joining wasn't inhibited
and JOIN-BLOCKS just updated the reference. Let's see if something
breaks and we need to keep track of references again. The reason I
gutted this is because preventing block joining = more ir blocks =
slower compiler, less optimizations.