Fix let and assignment conversion inside deleted lambdas.
Lets and assignments should not derive as non-returning, as that
doesn't make sense conceptually and causes blocks to become
unreachable incorrectly. This is a better solution than what was done
before for conversions inside deleted lambdas, as the earlier fix
doesn't generalize to assignment lambdas, which are also more prone to
this issue, since multiple calls may stick around to the assignment
lambda and derive as not returning, causing live code to get marked as
unreachable.

Fixes lp#1951887.

Also add a test case for a previous commit, which fixes lp#1951889.