Discourage assigning directly to SYMBOL-%INFO
The only uses of a setter were for constant coalescing (and the allocator,
which doesn't count), but coalescing the slot value would be wrong,
because it is unspecified how (SETF GET) alters a plist in terms of
destructive modification. We don't want sharing of plists that become EQ
by accident. SYMBOL-DBINFO is unlikely be EQUAL to that of another symbol,
so it's no real loss there.