Simplify FINISH-SYMBOLS
This got easier with the advent of the XC-STRICT-CL package, because we need
not concern ourselves with dozens of symbols that used to get shadowing-imported
into every SB-* package so that it could use COMMON-LISP but have easy access
to some things in SB-XC: without a package qualifier.
Now that all packages use XC-STRICT-CL which exports the right bunch of symbols
from the host's CL package, plus those from SB-XC as needed, anything that is
shadowed indicates that it should be a shadowing symbol in the target package.