Pin lockfree list node successor only if in *PINNED-OBJECTS*
This fixes a bug that the GC generation number of the current node must
not have any bearing on whether to pin NODE-%NEXT. i.e. if this node isn't
subject to copying (is in a generation higher than the evacuated one),
but the successor node is, then that node must be pinned.
And pin_object() was the wrong thing to call, as per its comment
"it is not OK ... if the object is not one which merits pinning"

Also this eliminates an extra tests on every pointer that implicitly
pinned something, so there is generally less work involved.