Don't recompute a slot, as well as moving a naming issue.
The NLX-INFO (formerly INFO, formerly NLX-INFO) slot of CLEANUP was
being recomputed in MAP-BLOCK-NLXes. Actually just use the slot,
allowing us to remove the poorly named function (which also handled
DX). We rename CLEANUP-INFO to CLEANUP-NLX-INFO as a matter of moving
the naming issue, and because the slot should really only hold
NLX-INFOs. Soon, this overloadedness will go away.