Generate PPC64ELF ABIv1 alien callbacks
In alien-callback-assembler-wrapper, look up
callback_wrapper_trampoline's address and table-of-contents entry from
its function descriptor.

Also, start the generated segment with a function descriptor that the
caller can use.  This descriptor points just past the end of itself to
the generated assembler wrapper.

With this change callback.impure passes on ppc64.