Enforce fewer side-effects of "pure" regression tests
- Capture before/after state of globaldb and global symbols for comparison.
- Do a better job clobbering remants of defining forms after each test file.
- Allow DEFSTRUCT and DEFTYPE in pure tests.

The intent of this change is to  get a larger set of tests that don't fork
a new child, which will provide more metrics on the effectiveness of GC
during the regression run.

Contrary to the comments, DEFUN and DEFMACRO do not render a test impure,
because DELETE-PACKAGE undoes those. Apparently thread creation is treated
as pure, evidenced by the mere existence of 'threads.pure.lisp'.
Also, creating a file is not impure, because state change outside the process
is usually irrelevant. A failing pure test may have side-effects by accident,
which isn't great, but we can live with it.