Remove scav_fun_header
The ifdef surrounding the function definition was from 4fa6bd72 presumably
to eliminate a C compiler warning about an unused static function.
The proper fix would have been to remove the definition and use.
But I did some archaeology to see why this was ever specific to some
architectures. The original condition use in scavtab[] was from 9f10bc10.
It may have been right or wrong back then, but is certainly wrong now
as the entire point is that you can't call it.