Add another ECL workaround
As long as the math journal is never written to, there is no need
to output a printed representation of any instance of SB-XC:FLOAT.
If for some reason the implementation decides to print one
(for internal reasons?) the default #S syntax is enough.

I don't know if this actually fixes lp#1956899 but one fewer method
clearly can't be wrong. I did the same with SBCL as host, which worked.