Share some code between trivial defmethod and real defmethod.
* Define the helpers early enough so that we can use them to also help
the expansion of trivial defmethod.
* Use the new generalized machinery to extend trivial defmethod to
handle any number of qualifiers, since it actually simplifies the code
and works just as well. (They still only take effect when the real
methods become a thing though.)
* No more incomplete and hacked up parsing, especially for special
cased names (the list of which is incomplete). The assertions don't
give helpful messages and we'll get a more helpful error when the real
methods get defined anyway.