sb-gmp: Define original functions more efficiently and portably.
Setting the fdefinition directly rather than going through a DEFUN and
APPLY is going to be more efficient. The old code also relied on
unportable behavior: LOAD-TIME-VALUE might be evaluated before the
DEFVAR is executed at load time. SBCL currently does not evaluate in
that order, but it may in the future (as some other implementations
do).