Fix standalone ldb some more
Different C compiler flags can add padding to the end of ucontext_t for the
same libc and kernel. Registers are always at the same offsets, so we just
need to write and read the same amounts. This fixes a load_gc_crashdump()
bug when compiled with stack smashing protection if the lisp that produced
the crashdump was not.