Disallow failure from COMPILE-FILE in cload tests
Impure tests convey status only through 'test-status.lisp-expr' , but nothing
wrote to that if the compile fails in such a way as to leave a do-nothing fasl.
Pure tests may at one point have aborted properly, but PURE-RUNNER is
complicated enough that it's hard to tell any more what used to work.
Anyway the fix is the same for both cases- explicitly note in *FAILURES*
that COMPILE-FILE says it failed.
And of course fix all the tests for which that used to happen.