Coalesce specialized arrays in the fasl dumper.
Don't rely on the IR1 namespace to do so: At most, we can rely on the
IR1 namespace scoping to enclose the compilation of a component, not a
whole file like the fasl dumper will do.

We still need the IR1 namespace similar table to coalesce similar
entries in the constant vector of code components; the fasl dumper is
not able to do that. But the fasl dumper is able to coalesce any
similar constant to be loaded as long as they are all in the same
file.