Remove OPAQUE-BOX and friends.
This partially reverts 369a4c8cf347d10cd37e878915e54dfaeeb023e7. It is
actually possible to do what :IGNORE-IT used to do in totally plain
and conforming Lisp code, by having a null load form. Thus we bring
back :IGNORE-IT but only as a convenient macro shortcut, not as the
magic non-conformant hack it used to be. A null load form would not
have actually worked in the past, because of the ancient (going as far
back as the introduction of make-load-form support in CMU CL) bug
fixed in 71e0191458d8ebfb433278c92d731bb421910884. What OPAQUE-BOX
really accomplished was just preventing the literal processing code to
recurse through the substructure, but we never should have done that
in the first place outside of the MAKE-LOAD-FORM-SAVING-SLOTS =>
DUMP-STRUCTURE case.  (Or when the creation and init forms actually
have subpart references.)

Advantages of the "new" scheme include:
* Localizing the structures that shouldn't be dumped to the make load
form method, which makes sense.
* Decreased maintenance burden from having to figure out how to thread
OPAQUE-BOX and OPAQUELY-QUOTE, when it is much easier to just declare
that a structure doesn't need to be grovelled or dumped.
* Possible extension to structures which are not as amenable to opaque
quoting, thereby allowing potentially more DEF!STRUCTs to get removed.