gencgc: Never use *ALLOCATION-POINTER*
There's nothing wrong with dynamic_space_free_pointer, other than they're
both about to be removed. 'next_free_page' is the high water mark in
dynamic space; you can compute an address from it as needed (not done yet).

And thus one of the variations of allocptr-*.inc goes the way of the dodo.