Preserve EQness of constants when block compiling.
* Use a dedicated FOP to dump named constant references. We used to
dump a FOP-FUNCALL, but we need to do something more exotic to support
forward references in both cold load and the target. The backpatching
is crucially needed to make cold load work, but it also allows us to
actually install constants into code components at load time via
%DEFCONSTANT, which solves the issues with DEFCONSTANT and block
compilation mixing. See the test case, which has now been fixed.

* Loading a dedicated FOP is faster than loading a dumped load time
lambda, not to mention smaller, but we are actually mostly using it to
help with the backpatching info. As a side effect, we can get rid of
all the bookkeeping for named constants in the dumper, since we now do
it by name. We can't even cache the result in the dumper because we
need to record backpatches into different code components.

* We get rid of the extremely ugly *HAIRY-DEFCONSTANTS* hack that was
only used to minimize the cost of block compiling DEFCONSTANTs and
reduce the amount of non-EQness of DEFCONSTANT in that context.

* Define a new info type to keep track of the forward references. For
the most common scenario, the info database isn't hit at all, since we
do things normally when the symbol is bound at the point of fasl load.

* A sort-of weird but justifiable side effect of this change: If you
compile

(defun f () foo)

where foo has been declared a constant, and load the fasl without
defining foo, (f) returns <unbound>, rather than throwing a load time
error. This matches DEFGLOBAL or ALWAYS-BOUND behavior, which is
totally acceptable and conforming for DEFCONSTANT, since DEFCONSTANT
is a promise that the constant will actually be defined and bound to
its real value in a reconstructed image via loading.