Rewrite a non-portable loop construct
The contortion here was to perform an iteration over the list
designator, without consing a list in the case that the designator was
an atom.  That's laudable, but in the context of the disassembler I
can't convince myself that it's worth avoiding the double-word of
consing, so rewrite it to use ENSURE-LIST and a straightforward
iteration.