Fix (trace foo :wherein bar :encapsulate t)
When tracing a function FOO via encapsulation, TRACE-WHEREIN-P is
invoked before entering FOO and in that case the top frame is FOO's
caller. So, don't skip the frame.

When tracing via breakpoints, TRACE-WHEREIN-P is invoked after
entering FOO in which case we want to continue skipping the top frame,
presumably to do the right thing with (trace foo :wherein foo).