Add a github CI job for xc fasl comparison
The xc fasls depend, through their debug info, on the mtime of the
source files.  This might change in the future, moving to e.g. some
content-based file id (to allow tools to determine whether a fasl is
up-to-date with respect to its source), but for now it is what it is.
This is awkward from the point of view of reproducibility of a given
git version, however, because `git clone` (and other git commands) do
not produce files with deterministic mtimes; instead, the mtimes are
derived from the current time -- understandably, given that many build
systems use timestamps to determine up-to-dateness.

In the particular case of a clean build from continuous integration,
though, we can restore potential reproducibility by modifying the
mtimes of a file in a clean checkout to that of the latest commit
modifying that file.  If we do that to the checkout about to be built
by ccl, and likewise a checkout about to be built by sbcl (in a new
job), then we can compare the xc fasls from each with the expectation
that, at least, those fasls don't differ because of source file
mtimes.