Simplify uncrossing machinery.
It looks like uncrossing used to be used for entire programs in the
past. But now we have much tighter control of where we want to
uncross, and it's overwhelmingly with symbols, and we really don't
need to worry about atoms hiding uncrossable symbols. Therefore,
simplify and remove some obsolete comments.

This also gives a noticeable cross-compile speedup, since it's used
quite a lot.

Also remove 2 uncross calls that were just stand-in's for testing the
symbol package.