Introduce more fine-grained integer simd pack types.
Split the type (simd-pack[-256] integer) into (simd-pack[-256] (A B)),
where A is one of {signed-byte,unsigned-byte} and where B is one of
{8,16,32,64}.  The benefits of doing so are more precise static type
checking, and a more helpful printed representation of integer SIMD packs.

Additionally, introduce a simd-pack[-256]-any primitive type that can be
used to implement VOPs whose implementation is independent of the actual
SIMD pack type, such as %simd-pack-low or %simd-pack-256-0.  This way, we
can get rid of the clumsy sb-vm:simd-pack[-256]-dispatch macros that lead
to unnecessary code bloat.