Add better hash mixer for fixnums
It is desirable that hashing of dynamic-space addresses work well
from lisp, but it did not because SXHASH randomizes very few bits
if addresses are near each other for some definition of "near".
And it's hard to know which bits are the nicely hashed ones.
So add a new choice of hash function for system-internal use.
For what it's worth, 32-bit wasn't as bad as 64-bit.