x86-64: Go back to using SIMD-PACK-DISPATCH.
The "any" types introduced too much type erasure when boxing and
unboxing, notably when spilling SIMD packs, resulting in SIMD values
"losing" their types, as unboxed values on the stack do not keep track
of their type.

Also remove the PTYPE hack once and for all by splitting up the boxing
VOPs to specializie on the arg type. This should keep representation
selection fast since we only need to dispatch based on the ptype for
the integer AVX2 and SSE SCs.