Define FUNCTION-LAMBDA-EXPRESSION earlier
Now that we store a lambda expression in the LEXENV, as opposed to
computing one on demand, conversion of interpreter environments must
generate such a lambda when converting their environments to a
compiler LEXENV; the easiest way for them to do that is to call
FUNCTION-LAMBDA-EXPRESSION themselves.  This means that it needs to be
defined early enough to support uses of MACROLET in warm init.