Funnel COMPILE calls in CLOS through a wrapper function
This simplifies rebinding of the warning and note muffler, and allows
tracing or otherwise intercepting system-generated compilation.

Also: gensym is considered harmful. I'm noticing tons of compilations
to the identical lambda, and it's a lot easier to figure out why
without having to think about symbols that differ for no good reason.
(I suspect that if N threads all need to invoke a GF for the first time
at about the same time, they'll all cache miss, and then all compile
the identical dfun. This is horrible and is need of good solution)