Record whether inheritance is allowed when saving symbols.
When delaying toplevel form execution until after component loading in
the loader, shadowing would not effect the package environment
correctly during fasloading. Fix this by recording whether inheritance
for the symbol is allowed at the point of dumping. We are allowed to
do this because the specification says that in conforming code, the
package environment during compilation should be congruent to the
package environment during fasload.

This is the most expedient fix, but this could probably be done better
in a way that unifies the accessibility logic in cold load.