Fix really weird source of cross-host discrepancy.
When block compiling this file, using unquoting rather than read eval
caused some unintuitive list* fops to get dumped on sbcl host, whereas
ccl host wouldn't have used a list* fop. With the reader eval, sbcl
dumps the list structure similarly to CCL. It's so weird that we have
a host sensitivity here, since the dump logic should be host agnostic
at this point, but I suspect it has something to do with how this list
is literally spliced in into the next DEFTRANSFORM, causing
coalescing, and maybe because it was spliced into a host only function
in the same file. ?????????????