Don't try to fold named constant references inside backquote anymore
The gain for doing so was fairly marginal. The thing that really
mattered with the change adding the folding was telling the difference
between a named constant and anonymous constant inside the backq
constant folder, for the sake of sane load form semantics.

Not backq-constant-folding named constant references removes a
load-form hack which also just doesn't work with deferred TLF
processing, as the load form may get evaluated for value before the
constant definition for the named constant is evaluated.